Actualizacion en tabla publicacion_grado

+++ scriptCreacionInicial.sql
@@ -55,7 +55,7 @@ CREATE TABLE ESKHERE.[Publicacion_Estado](
 CREATE TABLE ESKHERE.[Publicacion_Grado](
 	[ID] [int] NOT NULL PRIMARY KEY IDENTITY(1,1),
 	[Descripcion] [nvarchar](255) NULL,
-	[Comision] int
+	[Comision] numeric(18,2)
 	);
 
 CREATE TABLE ESKHERE.[Usuario](
